Safari-609.1.11.1/LayoutTests/http/tests/navigation/post-goback1.html

<script>
  sessionStorage.formTargetShouldNavAndGoBack = true;
  if (window.testRunner) {
      testRunner.dumpAsText();
      testRunner.waitUntilDone();
      testRunner.clearBackForwardList();
      testRunner.dumpBackForwardList();
  }
  window.location.href = 'resources/page-that-posts.html';
</script>  

<p>Tests that POST-ing a page, navigating to another page and then going back is handled correctly. If this test passes, you will see a page indicating a form was POSTed.</p>
