LayoutTests/platform/ios-sim/http/tests/navigation/post-goback1-expected.txt

This page was requested with the HTTP method GET.

Parameters:



============== Back Forward List ==============
        http://127.0.0.1:8000/navigation/resources/page-that-posts.html  **nav target**
curr->  http://127.0.0.1:8000/navigation/resources/form-target.pl  **nav target**
        http://127.0.0.1:8000/navigation/resources/go-back.html  **nav target**
===============================================
